replace json server to mockapi and api errors processing

+++ src/components/ui/product-counter/ProductCounter.tsx
@@ -5,7 +5,8 @@ import Minus from '../../../assets/images/icons/minus.svg?react'
 import Plus from '../../../assets/images/icons/plus.svg?react'
 
 import { useChangeCountMutation, useDeleteFromCartMutation } from "../../../redux";
-import { FC } from 'react';
+import { FC, useEffect } from 'react';
+import useServerError from '../../../hooks/useServerError';
 
 interface ProductCounterProps {
   smallModifier?: boolean;
@@ -13,17 +14,38 @@ interface ProductCounterProps {
 }
 
 const ProductCounter: FC<ProductCounterProps> = ({ smallModifier, cardState }) => {
-  const [deleteFromCart] = useDeleteFromCartMutation()
-  const [changeCount] = useChangeCountMutation()
+  const [deleteFromCart, { isLoading: isDeleteLoading, isError: isDeleteError }] = useDeleteFromCartMutation()
+  const [changeCount, { isLoading: isChangeLoading, isError: isChangeError }] = useChangeCountMutation()
+  const {isTooManyRequestsError, doServerError} = useServerError()
+
   const changedId = cardState.mockid
+  /* spam timer */
+  useEffect(() => {
+    if (isChangeError || isDeleteError) {
+      doServerError();
+    }
+  }, [isChangeError, isDeleteError])
 
   async function handleChangeCount(symbol: ('-' | '+')): Promise<void> {
+    /* errors guard */
+    if (isTooManyRequestsError) {
+      alert('Too many requests to MockApi, await 20 seconds');
+      return
+    }
+
     let countChanger = cardState.count
     if (symbol === '+') {
       countChanger = cardState.count + 1
     } else if (symbol === '-') {
       if (countChanger === 1) {
-        await deleteFromCart({id: changedId}).unwrap()
+        /* spam server guard */
+        if (!isDeleteLoading) {
+          await deleteFromCart({id: changedId}).unwrap()
+        } else {
+          /* return to stop spaming to delete the same item */
+          return
+        }
+        /* return to stop function if item was delete*/
         return
       } else {
         countChanger = cardState.count - 1
@@ -34,9 +56,13 @@ const ProductCounter: FC<ProductCounterProps> = ({ smallModifier, cardState }) =
   
   return (
     <div className={!smallModifier ? styles["product-counter"] : `${styles["product-counter"]} ${styles["product-counter--small"]}`}>
-      <Minus onClick={() => handleChangeCount('-')} className={styles["product-counter__minus"]} width='14' height='14'/>
+      <button disabled={isChangeLoading} onClick={() => handleChangeCount('-')} className={styles["product-counter__button"]}>
+        <Minus  className={styles["product-counter__minus"]} width='14' height='14'/>
+      </button>
       <p className={styles["product-counter__count"]}>{cardState?.count}</p>
-      <Plus onClick={() => handleChangeCount('+')} className={styles["product-counter__plus"]} width='14' height='14'/>
+      <button disabled={isChangeLoading} onClick={() => handleChangeCount('+')} className={styles["product-counter__button"]}>
+        <Plus  className={styles["product-counter__plus"]} width='14' height='14'/>
+      </button>
     </div>
   )
 }
